commit
git-svn-id: http://svn.cern.ch/guest/lhcb/Gauss/trunk@7324 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 2b9ae7b3904b0cfcb0b9601a41d0dee4d12ffaa7

+++ Gauss/Sim/GaussRICH/src/RichAnalysis/RichG4ReconHpd.h
@@ -1,4 +1,4 @@
-// $Id: RichG4ReconHpd.h,v 1.2 2004-02-10 14:24:14 jonesc Exp $
+// $Id: RichG4ReconHpd.h,v 1.3 2005-06-16 11:39:59 seaso Exp $
 #ifndef RICHANALYSIS_RICHG4RECONHPD_H 
 #define RICHANALYSIS_RICHG4RECONHPD_H 1
 
@@ -45,8 +45,11 @@ private:
   double m_HpdPhCathodeRad;
   double m_HpdPixelXSize;
   double m_HpdPixelYSize;
-  double m_HpdPixelXNumCol;
-  double m_HpdPixelXNumRow;
+  int m_HpdPixelXNumCol;
+  int m_HpdPixelXNumRow;
+
+  //  double m_HpdPixelXNumCol;
+  // double m_HpdPixelXNumRow;
 
 };
